+++ docs/changelog.md
@@ -5,8 +5,15 @@
 ### Breaking Changes
 
 ### Bugfixes
+- "NFiniteSamplingInput supporting both trigger polarities via ConfigOption
 
 ### New Features
+- Re-introduced tilt correction (from old core) to the scanning probe toolchain.
+- Improved support for Stanford Research Systems signal generators
+- Expanded documentation of the microwave interface
+
+### Bugfixes
+- Old ODMR fits are now removed when starting a new measurement
 
 ### Other